+++ CHANGELOG.md
@@ -9,6 +9,25 @@
 - `Innmind\BlackBox\PHPUnit\Framework\Attributes\Group`
 - `Innmind\BlackBox\Set::via()`
 - `Innmind\BlackBox\Runner\Given::exclude()`
+- `Innmind\BlackBox\Prove`
+
+### Changed
+
+- The callable passed to `Application::tryToProve()` now has access to an instance of `Prove`
+
+### Depercated
+
+- `Innmind\BlackBox\Application::disableGlobalFunctions()`
+- `Innmind\BlackBox\Runner\proof()`
+- `Innmind\BlackBox\Runner\test()`
+- `Innmind\BlackBox\Runner\given()`
+- `Innmind\BlackBox\Runner\property()`
+- `Innmind\BlackBox\Runner\properties()`
+- `proof()`
+- `test()`
+- `given()`
+- `property()`
+- `properties()`
 
 ## 6.11.0 - 2026-03-18